remove editor
